requirements.txt
# 바이낸스 테스트넷 자동매매 봇 필수 라이브러리

# 바이낸스 API 클라이언트
python-binance==1.0.19

# 데이터 분석
pandas==2.1.4
numpy==1.26.2

# 환경 변수 관리
python-dotenv==1.0.0

# 추가 유용한 라이브러리 (선택사항)
# requests==2.31.0
# matplotlib==3.8.2  # 차트 그리기용


